+++ modules/rdns-forwarder/module.tf
@@ -3,86 +3,94 @@
 # Copyright (c) 2017 Board of Trustees University of Illinois
 
 terraform {
-  required_version = ">= 0.11"
+  required_version = ">= 0.12.9"
 
-  ## future (https://github.com/hashicorp/terraform/issues/16835)
-  #required_providers {
-  #  aws    = ">= 1.7"
-  #}
+  required_providers {
+    aws = ">= 2.32"
+  }
 }
 
 ## Inputs
 
 variable "instance_type" {
   description = "Type of EC2 instance to use for this RDNS Forwarder, e.g. t2.micro"
+  type        = string
 }
 
 variable "subnet_id" {
   description = "Subnet in which to launch, e.g. subnet-abcd1234.  Using a public-facing subnet is simplest, but a campus-facing or private-facing subnet will also work as long as it has outbound Internet connectivity (via a NAT Gateway)"
+  type        = string
 }
 
 # explicitly specifying private_ip helps minimize disruption to other services
 # when we destroy and re-create the RDNS Forwarder later on
 variable "private_ip" {
   description = "Private IPv4 address of this RDNS Forwarder (must be within the chosen subnet)"
+  type        = string
 }
 
 variable "associate_public_ip_address" {
   description = "set false if subnet is not public-facing"
+  type        = bool
   default     = true
 }
 
 variable "core_services_resolvers" {
   description = "IPv4 addresses of Core Services Resolvers to query for University domains"
-  type        = "list"
+  type        = list(string)
 }
 
 variable "zone_update_minute" {
   description = "Minute (between 0 and 59) to perform the hourly zone list update"
+  type        = string
   default     = "0"
 }
 
 variable "full_update_day_of_month" {
   description = "Day of month (between 1 and 28) to perform the monthly full update"
+  type        = string
   default     = "1"
 }
 
 variable "full_update_hour" {
   description = "Hour (between 0 and 23, note UTC) to perform the monthly full update"
+  type        = string
   default     = "8"
 }
 
 variable "full_update_minute" {
   description = "Minute (between 0 and 59) to perform the monthly full update"
+  type        = string
   default     = "17"
 }
 
 variable "tags" {
   description = "Optional custom tags for all taggable resources"
-  type        = "map"
+  type        = map
   default     = {}
 }
 
 variable "tags_instance" {
   description = "Optional custom tags for aws_instance resource"
-  type        = "map"
+  type        = map
   default     = {}
 }
 
 variable "tags_volume" {
   description = "Optional custom volume_tags for aws_instance resource"
-  type        = "map"
+  type        = map
   default     = {}
 }
 
 variable "tags_security_group" {
   description = "Optional custom tags for aws_security_group resource"
-  type        = "map"
+  type        = map
   default     = {}
 }
 
 variable "key_name" {
   description = "Optional Key Pair name (for SSH access)"
+  type        = string
   default     = ""
 }
 
@@ -100,25 +108,25 @@ variable "ansible_pull_checkout" {
 ## Outputs
 
 output "id" {
-  value = "${aws_instance.forwarder.id}"
+  value = aws_instance.forwarder.id
 }
 
 output "private_ip" {
-  value = "${aws_instance.forwarder.private_ip}"
+  value = aws_instance.forwarder.private_ip
 }
 
 output "security_group_id" {
-  value = "${aws_security_group.rdns.id}"
+  value = aws_security_group.rdns.id
 }
 
 ## Resources
 
 data "aws_subnet" "selected" {
-  id = "${var.subnet_id}"
+  id = var.subnet_id
 }
 
 data "aws_vpc" "selected" {
-  id = "${data.aws_subnet.selected.vpc_id}"
+  id = data.aws_subnet.selected.vpc_id
 }
 
 # Get the latest Amazon Linux AMI named e.g. amzn-ami-hvm-2016.09.1.20170119-x86_64-gp2
@@ -133,39 +141,39 @@ data "aws_ami" "ami" {
 }
 
 data "template_file" "user_data" {
-  template = "${file("${path.module}/user_data.sh")}"
+  template = file("${path.module}/user_data.sh")
 
-  vars {
-    vpc_cidr   = "${data.aws_vpc.selected.cidr_block}"
-    amazon_dns = "${cidrhost(data.aws_vpc.selected.cidr_block,2)}"
+  vars = {
+    vpc_cidr   = data.aws_vpc.selected.cidr_block
+    amazon_dns = cidrhost(data.aws_vpc.selected.cidr_block,2)
 
     # format as YAML list
-    forwarders_list          = "${join("\n", formatlist("  - %s", var.core_services_resolvers))}"
+    forwarders_list          = join("\n", formatlist("  - %s", var.core_services_resolvers))
     ansible_logfile          = "/var/log/ansible.log"
-    ansible_pull_url         = "${var.ansible_pull_url}"
-    ansible_pull_checkout    = "${var.ansible_pull_checkout}"
+    ansible_pull_url         = var.ansible_pull_url
+    ansible_pull_checkout    = var.ansible_pull_checkout
     ansible_pull_directory   = "/root/aws-enterprise-vpc"
-    zone_update_minute       = "${var.zone_update_minute}"
-    full_update_day_of_month = "${var.full_update_day_of_month}"
-    full_update_hour         = "${var.full_update_hour}"
-    full_update_minute       = "${var.full_update_minute}"
+    zone_update_minute       = var.zone_update_minute
+    full_update_day_of_month = var.full_update_day_of_month
+    full_update_hour         = var.full_update_hour
+    full_update_minute       = var.full_update_minute
   }
 }
 
 # EC2 instance
 
 resource "aws_instance" "forwarder" {
-  tags                        = "${merge(var.tags, var.tags_instance)}"
-  volume_tags                 = "${merge(var.tags, var.tags_volume)}"
-  ami                         = "${data.aws_ami.ami.id}"
-  instance_type               = "${var.instance_type}"
-  subnet_id                   = "${var.subnet_id}"
-  private_ip                  = "${var.private_ip}"
-  associate_public_ip_address = "${var.associate_public_ip_address}"
-  key_name                    = "${var.key_name}"
-  vpc_security_group_ids      = ["${aws_security_group.rdns.id}"]
-  iam_instance_profile        = "${aws_iam_instance_profile.instance_profile.name}"
-  user_data                   = "${data.template_file.user_data.rendered}"
+  tags                        = merge(var.tags, var.tags_instance)
+  volume_tags                 = merge(var.tags, var.tags_volume)
+  ami                         = data.aws_ami.ami.id
+  instance_type               = var.instance_type
+  subnet_id                   = var.subnet_id
+  private_ip                  = var.private_ip
+  associate_public_ip_address = var.associate_public_ip_address
+  key_name                    = var.key_name
+  vpc_security_group_ids      = [aws_security_group.rdns.id]
+  iam_instance_profile        = aws_iam_instance_profile.instance_profile.name
+  user_data                   = data.template_file.user_data.rendered
 
   lifecycle {
     # Avoids unnecessary destruction and recreation of the instance by Terraform
@@ -179,9 +187,9 @@ resource "aws_instance" "forwarder" {
 # Security Group
 
 resource "aws_security_group" "rdns" {
-  tags        = "${merge(var.tags, var.tags_security_group)}"
+  tags        = merge(var.tags, var.tags_security_group)
   name_prefix = "rdns-"
-  vpc_id      = "${data.aws_vpc.selected.id}"
+  vpc_id      = data.aws_vpc.selected.id
 
   lifecycle {
     create_before_destroy = true
@@ -190,7 +198,7 @@ resource "aws_security_group" "rdns" {
 
 resource "aws_security_group_rule" "allow_outbound" {
   # note: tags not supported
-  security_group_id = "${aws_security_group.rdns.id}"
+  security_group_id = aws_security_group.rdns.id
   type              = "egress"
   protocol          = "-1"
   from_port         = 0
@@ -200,40 +208,40 @@ resource "aws_security_group_rule" "allow_outbound" {
 
 resource "aws_security_group_rule" "allow_dns_udp" {
   # note: tags not supported
-  security_group_id = "${aws_security_group.rdns.id}"
+  security_group_id = aws_security_group.rdns.id
   type              = "ingress"
   from_port         = 53
   to_port           = 53
   protocol          = "udp"
-  cidr_blocks       = ["${data.aws_vpc.selected.cidr_block}"]
+  cidr_blocks       = [data.aws_vpc.selected.cidr_block]
 }
 
 resource "aws_security_group_rule" "allow_dns_tcp" {
   # note: tags not supported
-  security_group_id = "${aws_security_group.rdns.id}"
+  security_group_id = aws_security_group.rdns.id
   type              = "ingress"
   from_port         = 53
   to_port           = 53
   protocol          = "tcp"
-  cidr_blocks       = ["${data.aws_vpc.selected.cidr_block}"]
+  cidr_blocks       = [data.aws_vpc.selected.cidr_block]
 }
 
 resource "aws_security_group_rule" "allow_icmp" {
   # note: tags not supported
-  security_group_id = "${aws_security_group.rdns.id}"
+  security_group_id = aws_security_group.rdns.id
   type              = "ingress"
   from_port         = "-1"                                    # ICMP type number
   to_port           = "-1"                                    # ICMP code
   protocol          = "icmp"
-  cidr_blocks       = ["${data.aws_vpc.selected.cidr_block}"]
+  cidr_blocks       = [data.aws_vpc.selected.cidr_block]
 }
 
 # IAM Role
 
 resource "aws_iam_instance_profile" "instance_profile" {
   # note: tags not supported
-  name = "${aws_iam_role.role.name}"
-  role = "${aws_iam_role.role.name}"
+  name = aws_iam_role.role.name
+  role = aws_iam_role.role.name
 }
 
 resource "aws_iam_role" "role" {
@@ -262,7 +270,7 @@ EOF
 resource "aws_iam_role_policy" "inline1" {
   # note: tags not supported
   name_prefix = "rdns-forwarder-"
-  role        = "${aws_iam_role.role.name}"
+  role        = aws_iam_role.role.name
 
   policy = <<EOF
 {