+++ modules/vpn-connection/module.tf
@@ -3,30 +3,33 @@
 # Copyright (c) 2017 Board of Trustees University of Illinois
 
 terraform {
-  required_version = ">= 0.11"
+  required_version = ">= 0.12.9"
 
-  ## future (https://github.com/hashicorp/terraform/issues/16835)
-  #required_providers {
-  #  aws    = ">= 1.7"
-  #}
+  required_providers {
+    aws = ">= 2.32"
+  }
 }
 
 ## Inputs
 
 variable "name" {
   description = "tag:Name for this VPN Connection"
+  type        = string
 }
 
 variable "vpn_gateway_id" {
   description = "VPN Gateway to use for this VPN connection, e.g. vgw-abcd1234"
+  type        = string
 }
 
 variable "customer_gateway_id" {
   description = "Customer Gateway to connect to, e.g. cgw-abcd1234"
+  type        = string
 }
 
 variable "create_alarm" {
   description = "Set true to create a CloudWatch Metric Alarm"
+  type        = bool
   default     = false
 }
 
@@ -37,24 +40,25 @@ variable "create_alarm" {
 # remaining tunnel is still up and the VPN Connection is still functional.
 variable "alarm_requires_both_tunnels" {
   description = "Set true if *both* tunnels must be up for the alarm state to be OK"
+  type        = bool
   default     = false
 }
 
 variable "alarm_actions" {
   description = "Optional list of actions (ARNs) to execute when the alarm transitions into an ALARM state from any other state, e.g. [arn:aws:sns:us-east-2:999999999999:vpn-monitor-topic]"
-  type        = "list"
+  type        = list(string)
   default     = []
 }
 
 variable "insufficient_data_actions" {
   description = "Optional list of actions (ARNs) to execute when the alarm transitions into an INSUFFICIENT_DATA state from any other state."
-  type        = "list"
+  type        = list(string)
   default     = []
 }
 
 variable "ok_actions" {
   description = "Optional list of actions (ARNs) to execute when the alarm transitions into an OK state from any other state."
-  type        = "list"
+  type        = list(string)
   default     = []
 }
 
@@ -65,25 +69,25 @@ provider "aws" {
 
 variable "tags" {
   description = "Optional custom tags for all taggable resources"
-  type        = "map"
+  type        = map
   default     = {}
 }
 
 variable "tags_vpn_connection" {
   description = "Optional custom tags for aws_vpn_connection resource"
-  type        = "map"
+  type        = map
   default     = {}
 }
 
 ## Outputs
 
 output "id" {
-  value = "${aws_vpn_connection.vpn.id}"
+  value = aws_vpn_connection.vpn.id
 }
 
 output "customer_gateway_configuration" {
   sensitive = true
-  value     = "${aws_vpn_connection.vpn.customer_gateway_configuration}"
+  value     = aws_vpn_connection.vpn.customer_gateway_configuration
 }
 
 # wrapped in here-document delimiters for parsing convenience
@@ -96,19 +100,17 @@ output "customer_gateway_configuration_heredoc" {
 
 # Get region of default provider
 
-data "aws_region" "current" {
-  current = true
-}
+data "aws_region" "current" {}
 
 # VPN Connection
 
 resource "aws_vpn_connection" "vpn" {
-  tags = "${merge(var.tags, map(
-    "Name", var.name,
-  ), var.tags_vpn_connection)}"
+  tags = merge(var.tags, {
+    Name = var.name
+  }, var.tags_vpn_connection)
 
-  vpn_gateway_id      = "${var.vpn_gateway_id}"
-  customer_gateway_id = "${var.customer_gateway_id}"
+  vpn_gateway_id      = var.vpn_gateway_id
+  customer_gateway_id = var.customer_gateway_id
   type                = "ipsec.1"
 }
 
@@ -116,36 +118,35 @@ resource "aws_vpn_connection" "vpn" {
 # https://docs.aws.amazon.com/solutions/latest/vpn-monitor/
 
 locals {
-  alarm_threshold   = "${var.alarm_requires_both_tunnels ? 2 : 1}"
+  alarm_threshold   = var.alarm_requires_both_tunnels ? 2 : 1
   alarm_description = "verify that ${var.alarm_requires_both_tunnels ? "both tunnels are" : "at least one tunnel is"} UP"
 }
 
 resource "aws_cloudwatch_metric_alarm" "vpnstatus" {
   # note: tags not supported
-  provider          = "aws.vpn_monitor"
-  count             = "${var.create_alarm ? 1 : 0}"
+  provider          = aws.vpn_monitor
+  count             = var.create_alarm ? 1 : 0
   alarm_name        = "${aws_vpn_connection.vpn.id} | ${var.name}"
-  alarm_description = "${local.alarm_description}"
+  alarm_description = local.alarm_description
   namespace         = "VPNStatus"
-  metric_name       = "${aws_vpn_connection.vpn.id}"
+  metric_name       = aws_vpn_connection.vpn.id
 
   dimensions = {
-    CGW = "${var.customer_gateway_id}"
-    VGW = "${var.vpn_gateway_id}"
+    CGW = var.customer_gateway_id
+    VGW = var.vpn_gateway_id
 
     # data value indicating region of VPN connection, which may be
     # different from region of CloudWatch Metric
-    Region = "${data.aws_region.current.name}"
+    Region = data.aws_region.current.name
   }
 
   statistic           = "Minimum"
   comparison_operator = "LessThanThreshold"
-  threshold           = "${local.alarm_threshold}"
+  threshold           = local.alarm_threshold
   period              = "300"
   evaluation_periods  = "2"
 
-  # note workaround for https://github.com/hashicorp/terraform/issues/11453
-  alarm_actions             = ["${var.alarm_actions}"]
-  insufficient_data_actions = ["${var.insufficient_data_actions}"]
-  ok_actions                = ["${var.ok_actions}"]
+  alarm_actions             = var.alarm_actions
+  insufficient_data_actions = var.insufficient_data_actions
+  ok_actions                = var.ok_actions
 }